# modules/control-panel-backups-api/openapi/v1/control-panel-backups.yaml
openapi: 3.1.0
info:
  title: Control Panel Backups API
  version: 1.0.0
paths:
...
  /api/v1/control-panel/backups/snapshots/{snapshot}/restore:
    post: {operationId: control.panel.backups.snapshots.restore, security: [{sanctum: []}], responses: {'202': {description: Snapshot restore queued}}}
...
components:
  securitySchemes:
    sanctum: {type: http, scheme: bearer, bearerFormat: Sanctum}